+++ src/features/products/ProductNavigator.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { Platform } from 'react-native';
 
+import { layout } from 'src/constants';
 import { RootStack } from 'src/navigation';
 import ProductDetailsScreen from './ProductDetailsScreen';
 
@@ -12,10 +12,7 @@ export default function renderProductNavigator() {
         component={ProductDetailsScreen}
         options={({ route }) => ({
           title: route.params.productName || 'Product',
-          headerTitleContainerStyle: {
-            width: Platform.OS === 'ios' ? '75%' : '90%',
-            alignItems: Platform.OS === 'ios' ? 'center' : 'flex-start',
-          },
+          headerTitleContainerStyle: layout.defaultHeaderTitleContainerStyle,
         })}
       />
     </RootStack.Group>